extended teleport bug fix to all teleports
There was a bug that caused teleports to fire multiple times if someone was jumping through them, similar to the bug for the teleport in trigger. I implemented the same bug fix on all the teleports. I also added back in the line that switch camera modes when leaving the lost woods